:recycle: refactor : 페이지 단위 컴포넌트 구조 정리
